Added data dir volume bind to db and airflow containers
The data dir is ignored by default. This dir is necessary in the db and
airflow containers to allow the copy method in postgres to work and for
the PostgresOperator to function correctly. On inital project start it
is necessary to create this directory first otherwise permission errors
can occur.