Extend base airflow image to install dbt
This commit extends the base airflow image to install dbt. This approach
was used instead of _PIP_ADDITIONAL_REQUIREMENTS because the later
method would cause pip to install the dep on every container up.
Installing using an extended image means it will be installed once when
the image is built.