Update schemas and rename sagerx_dbt
I thought it was unusual to call everything sagerx_dbt
when we don't call airflow sagerx_airflow.
I tried to align our schemas with dbt's recommendations
staging -> intermediate -> marts
but we may want to revisit this as it also eliminates
the flatfile schema that a lot of our SQL currently uses.
Also - added a custom schema macro that removes the
concatenation of the target default schema with the
custom schema (i.e. staging vs dev_staging).