Pluralize RxNorm models per dbt docs
See https://docs.getdbt.com/guides/best-practices/how-we-structure/2-staging
Plural. SQL, and particularly SQL in dbt,
should read as much like prose as we can achieve.
We want to lean into the broad clarity and
declarative nature of SQL when possible. As such,
unless there’s a single order in your orders table,
plural is the correct way to describe what is
in a table with multiple rows.