Consolidate pharmacy files into one
Rationale for this is that I would have to create a dbt model
to union these 6 tables together which would add another 30GB
of file space if materialized as a table. The downside to this
approach is it takes longer because the COPY statements are no
longer run in parallel. I'm sure there's a way to do this more
efficiently, but I'm just trying to minimize storage space
requirements and also reduce the need for unnecessary extra
tables.