XPath for Media
Created an XPath for Media that looks for ObservationMedia
and then grabs an image file name (if it exists - need to build
a test for if it exists to reduce the noise probably) and also
the entire text of the component.

Next step is to build a dbt staging model to RegEx the NDC
out of the <Text/> element since XPath doesn't natively
support that.