Create equivalency folder
Also moved the existing mart called
brand_products_with_related_ndcs here.

This mart was purpose built for a
client and I think it can be made to
be more generalizable, but don't have
a clear vision for that right now.

Either way, it belongs in more of an
equivalency category and will likely
have other equivalency-type marts
accompanying it soon.