Move inactive_ingredients to ndc
Even though inactive ingredients are
at the NDC9 "product" level, I consider
them more of an NDC-level mart.

It brings up the fundamental question of
NDC9 "products" vs the RxNorm product
that they represent. I'm most likely
going to stick with products representing
the RxNorm concept of a product.

This will be important to clarify in docs.